requirements.txt
python-docx
pymupdf
pandas
camelot-py
tiktoken
chromadb
sentence-transformers
rouge-score
pydantic
einops
openpyxl
langdetect>=1.0.9
deep-translator>=1.11.4
llama-cpp-python
streamlit>=1.25.0
requests>=2.31.0